Move session page to session package
