+++ app/src/main/java/com/example/composeapp/ui/activity/MainActivity.kt
@@ -3,7 +3,7 @@ package com.example.composeapp.ui.activity
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.example.composeapp.ui.page.SessionPage
+import com.example.composeapp.ui.page.session.SessionPage
 import com.example.composeapp.ui.theme.ComposeAppTheme
 
 class MainActivity : ComponentActivity() {
