+++ app/src/main/java/com/example/composeapp/ui/page/session/SessionPage.kt
@@ -1,4 +1,4 @@
-package com.example.composeapp.ui.page
+package com.example.composeapp.ui.page.session
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
